Add some checks that the plugins are downloadable for all supported platforms. (Before I fix windows arm for plugins - while buildbot does not contain them)
